+++ test/services/county-committee/index.test.js
@@ -116,97 +116,4 @@ describe("county-committee service", function() {
       });
     });
   });
-
-  // describe("archiving", function() {
-  //   it("can archive a county-committees", done => {
-  //     assert(mock_county_committee);
-  //     assert(mock_county_committee.id);
-
-  //     ccService
-  //       .archive(mock_county_committee)
-  //       .then(archived_county_committee => {
-  //         assert(archived_county_committee._id === mock_county_committee._id);
-
-  //         const cc = CountyCommittee.findOne({
-  //           _id: mock_county_committee._id
-  //         }).then(cc => {
-  //           assert(!cc);
-
-  //           const CountyCommitteeArchive = mongoose.model(
-  //             "county-committee-archive"
-  //           );
-
-  //           CountyCommitteeArchive.findOne({
-  //             _id: mongoose.Types.ObjectId(mock_county_committee.id)
-  //           }).then(archived_cc => {
-  //             assert.deepEqual(archived_cc._id, mock_county_committee._id);
-
-  //             // Cleanup
-  //             done();
-  //           });
-  //         });
-  //       });
-  //   });
-  //   it("can un-archive a county-committees", done => {
-  //     const CountyCommitteeArchive = mongoose.model("county-committee-archive");
-
-  //     ccService.unArchive(mock_county_committee).then(county_committee => {
-  //       assert(mock_county_committee._id === county_committee._id);
-
-  //       const cc = CountyCommitteeArchive.findOne({
-  //         _id: county_committee._id
-  //       }).then(cc => {
-  //         assert(!cc);
-
-  //         const archived_cc = CountyCommittee.findOne({
-  //           _id: county_committee._id
-  //         }).then(unarchived_cc => {
-  //           assert.deepEqual(unarchived_cc._id, mock_county_committee._id);
-  //           done();
-  //         });
-  //       });
-  //     });
-  //   });
-
-  //   it("can archive county-committee members", done => {
-  //     const CountyCommitteeArchive = mongoose.model("county-committee-archive");
-  //     ccService.archive(mock_county_committee).then(function(county_committee) {
-  //       assert(county_committee);
-  //       assert(county_committee._id);
-  //       debugger;
-  //       ccService
-  //         .archiveCountyCommitteeMembers(county_committee)
-  //         .then(writeOpResult => {
-  //           assert(writeOpResult.deletedCount, 4590);
-  //           done();
-  //         });
-  //     });
-  //   });
-
-  //   it("can unarchive county-committee members", done => {
-  //     const ccService = app.service(app.get("apiPath") + "/county-committee");
-  //     const CountyCommitteeArchive = mongoose.model("county-committee-archive");
-
-  //     ccService
-  //       .unArchiveCountyCommitteeMembers(mock_county_committee._id)
-  //       .then(writeOpResult => {
-  //         assert(county_committee._id === mock_county_committee._id);
-  //         done();
-  //       });
-  //   });
-
-  //   after(function(done) {
-  //     const CountyCommitteeArchive = mongoose.model("county-committee-archive");
-  //     CountyCommitteeArchive.findOne({ _id: mock_county_committee._id }).then(
-  //       function(archived_county_committee) {
-  //         if (archived_county_committee) {
-  //           console.log(
-  //             "Looks like something didn't get cleaned up quite right. Deleting test archive data."
-  //           );
-  //           archived_county_committee.remove();
-  //         }
-  //       }
-  //     );
-  //   });
-  // });
 });
